Changed folder structure to something more familiar for the team
