+++ src/AppStyle.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const TempPhoneContainer = styled.section`
+  max-width: 480px;
+  margin: 0 auto;
+`;
+
+export const LayoutContainer = styled.section`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  box-sizing: border-box;
+`;
