+++ src/pages/Signup/SignupStyle.js
@@ -1,0 +1,8 @@
+import { TextField } from "@mui/material";
+import styled from "styled-components";
+
+export const StyledInputTextField = styled(TextField)`
+  && {
+    margin-bottom: 32px;
+  }
+`;
